+++ apostello/logs.py
@@ -3,14 +3,37 @@ import logging
 from django.conf import settings
 from twilio.base.exceptions import TwilioRestException
 
-from apostello.models import Keyword, Recipient, SmsInbound, SmsOutbound
-from apostello.twilio import twilio_client
+from .models import Keyword, Recipient, SmsInbound, SmsOutbound
+from .twilio import twilio_client
+from site_config.models import SiteConfiguration
 
 logger = logging.getLogger('apostello')
 
 
+def has_expired(datetime):
+    config = SiteConfiguration.get_solo()
+    if config.sms_expiration_date is None:
+        return False
+    else:
+        return datetime.date() < config.sms_expiration_date
+
+
+def cleanup_expired_sms():
+    """Remove expired messages."""
+    config = SiteConfiguration.get_solo()
+    if config.sms_expiration_date is None:
+        # no expiration, skip checks
+        return
+    for sms in SmsInbound.objects.filter(time_received__date__lt=config.sms_expiration_date):
+        sms.delete()
+    for sms in SmsOutbound.objects.filter(time_sent__date__lt=config.sms_expiration_date):
+        sms.delete()
+
+
 def handle_incoming_sms(msg):
     """Add incoming sms to log."""
+    if has_expired(msg.date_created):
+        return
     sms, created = SmsInbound.objects.get_or_create(sid=msg.sid)
     if created:
         sender, s_created = Recipient.objects.get_or_create(number=msg.from_)
@@ -32,6 +55,8 @@ def handle_incoming_sms(msg):
 
 def handle_outgoing_sms(msg):
     """Add outgoing sms to log."""
+    if has_expired(msg.date_sent):
+        return
     try:
         sms, created = SmsOutbound.objects.get_or_create(sid=msg.sid)
         if created: